Added fast add and shur product

+++ net/ActivationLayer.h
@@ -10,8 +10,8 @@
 
 class ActivationLayer : public Layer {
     public:
-        virtual double f(double x) = 0;
-        virtual double dfdx(double x) = 0;
+        virtual float f(float x) = 0;
+        virtual float dfdx(float x) = 0;
         Matrix* forward(Matrix* in) override;
         Matrix *activate(Matrix *in);
         Matrix *dactivate(Matrix *in);
