+++ net/layers/ReLU.h
@@ -10,8 +10,8 @@
 class ReLU : public ActivationLayer {
     public:
         ReLU();
-        double f(double x);
-        double dfdx(double x);
+        float f(float x);
+        float dfdx(float x);
 };
